gutread-agentic/requirements.txt
psycopg2-binary
dotenv
fastmcp
asyncio
crewai
crewai-tools[mcp]
crewai[google-genai]